Backend/email_otp/email_otp/email_otp/email_otp/email_otp_project/requirements.txt
Django>=3.2
djangorestframework>=3.12
google-auth>=2.0.0
djangorestframework-simplejwt>=4.7.2
